+++ shop/Models/Account.cs
@@ -14,16 +14,16 @@ namespace shop.Models
         }
 
         [Key]
-        public long AccountNumber { get; set; }
+        public int AccountNumber { get; set; }
         [Column(TypeName = "date")]
         public DateTime? CreateDate { get; set; }
-        public long? UserAdds { get; set; }
+        public int? UserAdds { get; set; }
         [Column("GroupID")]
-        public long? GroupId { get; set; }
+        public int? GroupId { get; set; }
         [Column("CustomerID")]
-        public long? CustomerId { get; set; }
+        public int? CustomerId { get; set; }
         [Column("SupplierID")]
-        public long? SupplierId { get; set; }
+        public int? SupplierId { get; set; }
         public bool? State { get; set; }
 
         [ForeignKey("CustomerId")]
